+++ Projects/Data/Model/Sources/GroupDTO.swift
@@ -6,8 +6,8 @@
 import Entity
 
 public struct GroupListItemResponseDTO: Decodable, Sendable {
-    public let groupId: Int64
-    public let meetingId: Int64
+    public let groupId: Int
+    public let meetingId: Int
     public let name: String
     public let themeTagCode: String
     public let themeTagDisplay: String
@@ -20,7 +20,7 @@ public struct GroupListItemResponseDTO: Decodable, Sendable {
 }
 
 public struct GroupMemberResponseDTO: Decodable, Sendable {
-    public let memberId: Int64
+    public let memberId: Int
     // TODO: 가입 미완료 테스트 계정으로 인한 임시 옵셔널 처리. 테스트 계정 정리 후 non-optional(String)로 복구
     public let nickname: String?
     public let profileImageUrl: String?
@@ -38,8 +38,8 @@ public struct CreateGroupRequestDTO: Encodable, Sendable {
 }
 
 public struct CreateGroupResponseDTO: Decodable, Sendable {
-    public let groupId: Int64
-    public let meetingId: Int64
+    public let groupId: Int
+    public let meetingId: Int
     public let name: String
     public let themeTagCode: String
 }
