+++ Projects/Data/Repository/Sources/GroupRepositoryImpl.swift
@@ -23,14 +23,14 @@ public final class GroupRepositoryImpl: GroupRepositoryProtocol {
         return response.toEntity()
     }
 
-    public func fetchGroupDetail(meetingId: Int64) async throws -> GroupDetail {
+    public func fetchGroupDetail(meetingId: Int) async throws -> GroupDetail {
         let response: GroupDetailResponseDTO = try await NetworkManager.shared.request(
             GroupEndpoint.fetchGroupDetail(meetingId: meetingId)
         )
         return response.toEntity()
     }
 
-    public func updateAttendance(groupId: Int64, attendanceStatus: AttendanceStatus) async throws {
+    public func updateAttendance(groupId: Int, attendanceStatus: AttendanceStatus) async throws {
         let requestDTO = UpdateAttendanceRequestDTO(attendanceStatus: attendanceStatus.rawValue)
         try await NetworkManager.shared.requestVoid(
             GroupEndpoint.updateAttendance(groupId: groupId, requestDTO)
