+++ Projects/Domain/DataInterface/Sources/VoteRepositoryProtocol.swift
@@ -6,6 +6,6 @@
 import Entity
 
 public protocol VoteRepositoryProtocol: Sendable {
-    func fetchDateVote(meetingId: Int64) async throws -> DateVote
-    func fetchPlaceVote(meetingId: Int64) async throws -> PlaceVote
+    func fetchDateVote(meetingId: Int) async throws -> DateVote
+    func fetchPlaceVote(meetingId: Int) async throws -> PlaceVote
 }
